Adding support for DSTs.
We have a couple of corner cases which included:
  * `&'a dyn`
  * `&str`
  * `&[T]`
These three cases are DSTs which are basically fat pointers, they use a
pointer to a vtable and the real pointer to the object, so they're
represented as a (usize, usize) when using the std::mem::size_of
function.

Also, added `memoffset` to dev-dependencies to have better testing.